requirements.txt
deprecation==2.1.0
img2pdf==0.5.1
lxml==5.2.2
packaging==24.1
pillow==10.4.0
PyPDF2==3.0.1
typing_extensions==4.12.2
pikepdf==5.6.1
